docker_main/requirements.txt
pandas
numpy
neo4j
python-dotenv
pyTelegramBotAPI
beautifulsoup4
requests
catboost
